requirements.txt
beautifulsoup4==4.4.1
Django==1.9.7
django-registration-redux==1.4
djangorestframework==3.3.3
Pillow==3.2.0
requests==2.10.0
wheel==0.24.0
